Move performance CLI command into commands module
