+++ src/cli/src/commands/index.ts
@@ -1,5 +1,6 @@
 export * from "./detect-test-regressions.js";
 export * from "./generate-feather-metadata.js";
 export * from "./generate-gml-identifiers.js";
+export * from "./performance.js";
 export * from "./memory.js";
 export * from "./watch.js";
